+++ src/GafferModule/PlugAlgoBinding.cpp
@@ -68,6 +68,18 @@ object findDestinationWrapper( Plug *plug, object predicate )
 	);
 }
 
+object findSourceWrapper( Plug *plug, object predicate )
+{
+	return PlugAlgo::findSource(
+		plug,
+		[&predicate] ( Plug *plug ) {
+			object o = predicate( PlugPtr( plug ) );
+			return o;
+		}
+	);
+}
+
+
 ValuePlugPtr createPlugFromData( const std::string &name, Plug::Direction direction, unsigned flags, const IECore::Data *value )
 {
 	IECorePython::ScopedGILRelease gilRelease;
@@ -132,6 +144,7 @@ void GafferModule::bindPlugAlgo()
 	def( "replacePlug", &replacePlug, ( arg( "parent" ), arg( "plug" ) ) );
 	def( "dependsOnCompute", &PlugAlgo::dependsOnCompute );
 	def( "findDestination", &findDestinationWrapper );
+	def( "findSource", &findSourceWrapper );
 
 	def( "createPlugFromData", &createPlugFromData );
 	def( "extractDataFromPlug", &extractDataFromPlug );